#49 "최근작업" 등록일은 종료 기간으로 올라가게 수정. 날짜타입 형식변경

+++ src/main/java/com/casker/portfolio/controller/PortfolioAdminController.java
@@ -107,7 +107,7 @@ public class PortfolioAdminController {
 	 * @return
 	 */
 	@RequestMapping(value = "/manegement/portfolio/editForm/{portfolioNo}", method = RequestMethod.GET)
-	public String editPortfolioForm(Model model, @PathVariable int portfolioNo) {
+	public String editPortfolioForm(Model model, @PathVariable long portfolioNo) {
 		
 		Portfolio portfolio = portfolioService.getPortfolioDetail(portfolioNo);
